% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_var_cut.R
\name{equal_var_cut}
\alias{equal_var_cut}
\title{"Equal variance cut"
Group observations so that the variance in each group is approximately constant.
(This is an internal function.)}
\usage{
equal_var_cut(pred_ord, groups, weights_ord)
}
\arguments{
\item{pred_ord}{See 'goodGLM'.}

\item{groups}{Number of groups.}

\item{weights_ord}{See 'goodGLM'.}
}
\value{

}
\description{
"Equal variance cut"
Group observations so that the variance in each group is approximately constant.
(This is an internal function.)
}
