% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hat_mat.R
\name{hat_mat}
\alias{hat_mat}
\title{Calculate the hat matrix
Calculate the hat matrix for a GLM object.
(This is an internal function.)}
\usage{
hat_mat(mod)
}
\arguments{
\item{mod}{}
}
\value{

}
\description{
Calculate the hat matrix
Calculate the hat matrix for a GLM object.
(This is an internal function.)
}
