% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_cov.R
\name{inverse_cov}
\alias{inverse_cov}
\title{"Inverse Covariance Matrix"
Calculate the matrix used in the quadratic form for the GOF test.
(This is an internal function.)}
\usage{
inverse_cov(mod, G)
}
\arguments{
\item{mod}{GLM object.}

\item{G}{See 'goodGLM'.}
}
\value{

}
\description{
"Inverse Covariance Matrix"
Calculate the matrix used in the quadratic form for the GOF test.
(This is an internal function.)
}
