% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodGLM.R
\name{goodGLM}
\alias{goodGLM}
\title{Generalized Hosmer-Lemeshow Test for GLMs
Performs a global goodness-of-fit test on objects of class 'glm'. The test is
based on a generalization of the Hosmer-Lemeshow test. For more information,
see the two papers by Surjanovic, Lockhart, and Loughin (2020) and
Surjanovic and Loughin (2021).}
\usage{
goodGLM(mod, groups, group_mode)
}
\arguments{
\item{mod}{GLM object.}

\item{groups}{Number of groups.}

\item{group_mode}{Group mode.}
}
\value{

}
\description{
Generalized Hosmer-Lemeshow Test for GLMs
Performs a global goodness-of-fit test on objects of class 'glm'. The test is
based on a generalization of the Hosmer-Lemeshow test. For more information,
see the two papers by Surjanovic, Lockhart, and Loughin (2020) and
Surjanovic and Loughin (2021).
}
